Fix how we join a transaction.
ITransaction no longer has a method register(), the intended way to
register a component in the current transaction seems to be join(). This
does not immediately fail due to the unqualified except clause that
hides this error, but the effect is that TM objects are no longer
correctly rolled back on transaction abort.