Fix linting issues
... which arise with the new flake8 version.

`tests` is no longer a valid flake8 param, as it is no top level
directory.

As it is a subfolder of the `src` folder, it gets linted anyway.

Also, `has_key` is deprecated and should be replaced by `in`.

modified:   tox.ini
modified:   src/Shared/DC/ZRDB/tests/testArgs.py