Fix the error output of SQL Methods
The error output uses __repr__, which in Zope2 was
inherited from SimpleItem.Item, which in turn pulls in
SimpleItem.PathReprProvider.

This inheritance pattern does not work in the same way
in Python3, losing PathReprProvider in DA. Adding it
there, at the very first position, yields the correct
output.

Test this with:
>>> from Products.ZSQLMethods.SQL import SQL
>>> x = SQL('a', 'a', 'a', '', 'select * from data')
>>> repr(x)
'<SQL at a>'